+++ src/Audit.EntityFramework/DbContextHelper.Net45.cs
@@ -117,13 +117,13 @@ namespace Audit.EntityFramework
             {
                 if (settings.OverrideProperties.ContainsKey(propName))
                 {
-                    // property overriden with a constant value
-                    value = settings.OverrideProperties[propName];
+                    // property overriden with a func value
+                    value = settings.OverrideProperties[propName].Invoke(entry);
                     return true;
                 }
                 if (settings.FormatProperties.ContainsKey(propName))
                 {
-                    // property overriden with a func value
+                    // property formatted
                     value = settings.FormatProperties[propName].Invoke(currentValue);
                     return true;
                 }