+++ test/Audit.EntityFramework.Core.UnitTest/InheritanceTests.cs
@@ -33,7 +33,7 @@ namespace Audit.EntityFramework.Full.UnitTest
             Audit.EntityFramework.Configuration.Setup()
                 .ForContext<DepartmentContext1>(_ => _
                     .IncludeEntityObjects()
-                    .ForEntity<Department>(emp => emp.Override(e => e.Name, "Override 1"))
+                    .ForEntity<Department>(emp => emp.Override(e => e.Name, e => "Override " + e.CurrentValues["Name"]))
                     .AuditEventType("{context}:Type1"));
 
             Audit.EntityFramework.Configuration.Setup()
@@ -77,7 +77,7 @@ namespace Audit.EntityFramework.Full.UnitTest
             Assert.AreEqual("DepartmentContext1:Type1", evs[0].EventType);
             Assert.AreEqual("DepartmentContext2:Type2", evs[1].EventType);
 
-            Assert.AreEqual("Override 1", evs[0].EntityFrameworkEvent.Entries[0].ColumnValues["Name"] as string);
+            Assert.AreEqual("Override test 1", evs[0].EntityFrameworkEvent.Entries[0].ColumnValues["Name"] as string);
             Assert.AreEqual("Override 2", evs[1].EntityFrameworkEvent.Entries[0].ColumnValues["Name"] as string);
             Assert.IsNotNull(evs[0].EntityFrameworkEvent.Entries[0].Entity);
             Assert.IsNull(evs[1].EntityFrameworkEvent.Entries[0].Entity);