+++ test/Audit.EntityFramework.UnitTest/Model.Context.cs
@@ -14,8 +14,7 @@ namespace Audit.EntityFramework.UnitTest
     using System.Data.Entity.Infrastructure;
     using System.Data.Entity.Core.Objects;
     using System.Linq;
-    using Core;
-
+    
     public partial class BlogsEntities : AuditDbContext
     {
         public BlogsEntities()