+++ test/Audit.EntityFramework.UnitTest/Post.cs
@@ -6,11 +6,12 @@
 //     Manual changes to this file will be overwritten if the code is regenerated.
 // </auto-generated>
 //------------------------------------------------------------------------------
+
 namespace Audit.EntityFramework.UnitTest
 {
     using System;
     using System.Collections.Generic;
-
+    
     public partial class Post
     {
         public int Id { get; set; }
@@ -18,7 +19,7 @@ namespace Audit.EntityFramework.UnitTest
         public System.DateTime DateCreated { get; set; }
         public string Content { get; set; }
         public int BlogId { get; set; }
-
+    
         public virtual Blog Blog { get; set; }
     }
 }